pyproject.toml
[project]
name = "learning-hub"
version = "0.1.0"
description = "learning-hub"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "django-debug-toolbar>=6.1.0",
    "django-markdownx>=4.0.9",
    "dotenv>=0.9.9",
    "ipython>=9.7.0",
    "markdown>=3.10",
    "pillow>=12.0.0",
]
